Add permission checking to permissions management

+++ app/Http/Controllers/Manage/PermissionsController.php
@@ -15,21 +15,29 @@ class PermissionsController extends Controller
     /**
      * Display a listing of the resource.
      *
-     * @return \Illuminate\Http\Response
+     * @return \Illuminate\Contracts\Foundation\Application|\Illuminate\Http\RedirectResponse|\Illuminate\Routing\Redirector
      */
     public function index()
     {
-        $perms  = DB::table('permissions')->orderBy('created_at', 'desc')->get();
-        return view('manage.permissions.index')->with('perms', $perms);
+        if (\Auth::user()->hasPermissionTo('Permissions.See')) {
+            $perms = DB::table('permissions')->orderBy('created_at', 'desc')->get();
+            return view('manage.permissions.index')->with('perms', $perms);
+        }else{
+            return redirect('/manage')->with('error','Unauthorized');
+        }
     }
     /**
      * Show the form for creating a new resource.
      *
-     * @return \Illuminate\Http\Response
+     * @return \Illuminate\Contracts\Foundation\Application|\Illuminate\Contracts\View\Factory|\Illuminate\View\View
      */
     public function create()
     {
-        return view('manage.permissions.create');
+        if (\Auth::user()->hasPermissionTo('Permissions.Create')) {
+            return view('manage.permissions.create');
+        }else{
+            return redirect('/manage/permissions')->with('error','Unauthorized');
+        }
     }
 
     /**
@@ -40,19 +48,23 @@ class PermissionsController extends Controller
      */
     public function store(Request $request)
     {
-        $validator =  Validator::make($request->all(), [
-            "perm_name" => "required|max:30|unique:permissions,name",
-            'api_perm'=>'nullable'
-        ]);
-        if ($validator->passes()) {
-            if(!$request->api_perm){
-                $permission = Permission::create(['name' => $request->perm_name]);
-            }else{
-                $permission = Permission::create(['name' => $request->perm_name,'guard_name' => 'api']);
+        if (\Auth::user()->hasPermissionTo('Permissions.Create')) {
+            $validator = Validator::make($request->all(), [
+                "perm_name" => "required|max:30|unique:permissions,name",
+                'api_perm' => 'nullable'
+            ]);
+            if ($validator->passes()) {
+                if (!$request->api_perm) {
+                    $permission = Permission::create(['name' => $request->perm_name]);
+                } else {
+                    $permission = Permission::create(['name' => $request->perm_name, 'guard_name' => 'api']);
+                }
+                return response()->json(['success' => 'The <b>' . $request->perm_name . '</b> permission has been created!']);
             }
-            return response()->json(['success' => 'The <b>'.$request->perm_name.'</b> permission has been created!']);
+            return response()->json($validator->errors());
+        }else{
+            return response()->json(['error' => 'Unauthorized']);
         }
-        return response()->json($validator->errors());
     }
 
 
@@ -64,8 +76,12 @@ class PermissionsController extends Controller
      */
     public function destroy($id)
     {
-        $perm = Permission::find($id);
-        $perm->delete();
-        return response()->json(['success' => 'The permission has successfully been deleted']);
+        if (\Auth::user()->hasPermissionTo('Permissions.Delete')) {
+            $perm = Permission::find($id);
+            $perm->delete();
+            return response()->json(['success' => 'The permission has successfully been deleted']);
+        }else{
+            return response()->json(['error' => 'Unauthorized']);
+        }
     }
 }
